live-control/requirements.txt
flask==3.0.3
mss==9.0.1
Pillow==10.4.0
pyautogui==0.9.54
pywin32==306
